requirements.txt
certifi==2018.11.29
chardet==3.0.4
configparser==3.5.0
Deprecated==1.2.4
idna==2.8
PyGithub==1.43.4
PyJWT==1.7.1
requests==2.21.0
urllib3==1.24.1
wrapt==1.10.11
